rename user-facing tools `foundry => kasmer`
