+++ README.md
@@ -50,11 +50,11 @@ To install `kmultiversx` and its dependencies into a virtual environment, run
 poetry -C kmultiversx install
 ```
 
-After the installation, the Python package `kmultiversx` and CLI tools `mandos` and `foundry` will be available via the `poetry run` command
+After the installation, the Python package `kmultiversx` and CLI tools `mandos` and `kasmer` will be available via the `poetry run` command
 
 ```
 poetry -C kmultiversx run mandos --help
-poetry -C kmultiversx run foundry --help
+poetry -C kmultiversx run kasmer --help
 ```
 
 Or you can activate the virtual environment managed by `poetry` and use the commands directly
@@ -70,7 +70,7 @@ Alternatively, you can install `kmultiversx` globally
 make -C kmultiversx build
 pip install kmultiversx/dist/*.whl
 mandos --help
-foundry --help
+kasmer --help
 ```
 
 ## Run
@@ -129,7 +129,7 @@ deps/wasm-semantics/elrond-config.md:324:10
 
 The coverage files generated by `krun` are located under the build directory (`.build/defn/llvm/mandos-kompiled`). To clean up the coverage data generated, run `make clean-coverage`.
 
-## Writing Foundry-like Parametric Tests
+## Writing Property Tests
 
 ### Create a test contract
 
@@ -168,9 +168,9 @@ let adder = testapi::deploy_contract(
 self.adder_address().set(&adder);
 ```
 
-Once you've implemented the init function, it's essential to make sure the test runner knows the correct file path to use when deploying the test contract. To achieve this, you'll need to create a `foundry.json` file in the root directory of your test contract. In this file, specify the relative path to the contract's WASM file that you want to deploy for testing.
+Once you've implemented the init function, it's essential to make sure the test runner knows the correct file path to use when deploying the test contract. To achieve this, you'll need to create a `kasmer.json` file in the root directory of your test contract. In this file, specify the relative path to the contract's WASM file that you want to deploy for testing.
 
-For example, assuming your test contract is named `test_adder`, and you want to deploy the `adder.wasm` contract located in the ../../../deps/mx-sdk-rs/contracts/examples/adder/output/ directory, your `foundry.json` file should look like this:
+For example, assuming your test contract is named `test_adder`, and you want to deploy the `adder.wasm` contract located in the ../../../deps/mx-sdk-rs/contracts/examples/adder/output/ directory, your `kasmer.json` file should look like this:
 
 ```json
 {
@@ -180,7 +180,7 @@ For example, assuming your test contract is named `test_adder`, and you want to 
 }
 ```
 
-By providing the correct file path in the `foundry.json` file, the test runner will be able to deploy the specified contract during the testing process. With the `init` function in place and the contract file path specified, you are now ready to write test cases in endpoints.
+By providing the correct file path in the `kasmer.json` file, the test runner will be able to deploy the specified contract during the testing process. With the `init` function in place and the contract file path specified, you are now ready to write test cases in endpoints.
 
 ### Writing test cases in endpoints
 
@@ -234,7 +234,7 @@ By combining pranks with storage access and assertions, we can comprehensively t
 
 To run the tests for your Smart Contracts, ensure that you have fulfilled the following prerequisites:
 
-1. Compile the semantics by executing `make build-foundry`.
+1. Compile the semantics by executing `make build-kasmer`.
 2. Add a `<path to adder contract>/multicontract.toml` file to the adder contract, something like:
     ```
     [settings]
@@ -260,12 +260,12 @@ mxpy contract build <path to test contract>
 ```
 
 
-Run the `foundry` tool with the test contract's path as the argument:
+Run the `kasmer` tool with the test contract's path as the argument:
 
 ```shell
-foundry --definition-dir .build/defn/llvm/foundry-kompiled --directory <path to test contract>
+kasmer --definition-dir .build/defn/llvm/kasmer-kompiled --directory <path to test contract>
 ```
 
-The `foundry` tool will deploy the test contract using the arguments specified in the `foundry.json` file located in the test directory. It will extract the names and argument types of the test endpoints. Subsequently, the script will test these endpoints using random inputs generated with the `hypothesis` library, enabling fuzz testing. If it encounters an input that falsifies the assertions made in the test cases, it attempts to shrink the input and identify a minimal failing example.
+The `kasmer` tool will deploy the test contract using the arguments specified in the `kasmer.json` file located in the test directory. It will extract the names and argument types of the test endpoints. Subsequently, the script will test these endpoints using random inputs generated with the `hypothesis` library, enabling fuzz testing. If it encounters an input that falsifies the assertions made in the test cases, it attempts to shrink the input and identify a minimal failing example.
 
 By following these steps, you can efficiently and comprehensively evaluate your Smart Contracts, ensuring their correctness and reliability in various scenarios and inputs.
