+++ kmultiversx/src/tests/conftest.py
@@ -7,20 +7,20 @@ from pyk.ktool.krun import KRun
 
 def pytest_addoption(parser: pytest.Parser) -> None:
     parser.addoption(
-        '--foundry-llvm-dir',
-        dest='foundry_llvm_dir',
+        '--kasmer-llvm-dir',
+        dest='kasmer_llvm_dir',
         type=dir_path,
-        help='Existing Foundry LLVM kompiled directory.',
+        help='Existing Kasmer LLVM kompiled directory.',
     )
 
 
 @pytest.fixture(scope='session')
-def foundry_llvm_dir(pytestconfig: pytest.Config) -> Path:
-    ldir = pytestconfig.getoption('foundry_llvm_dir')
+def kasmer_llvm_dir(pytestconfig: pytest.Config) -> Path:
+    ldir = pytestconfig.getoption('kasmer_llvm_dir')
     assert isinstance(ldir, Path)
     return ldir
 
 
 @pytest.fixture(scope='session')
-def foundry_llvm_krun(foundry_llvm_dir: Path) -> KRun:
-    return KRun(foundry_llvm_dir)
+def kasmer_llvm_krun(kasmer_llvm_dir: Path) -> KRun:
+    return KRun(kasmer_llvm_dir)
