+++ kmultiversx/src/tests/unit/test_generate_claim.py
@@ -4,19 +4,19 @@ from typing import TYPE_CHECKING
 
 from pyk.kast.manip import split_config_from
 
-from kmultiversx.foundry import generate_claim
+from kmultiversx.kasmer import generate_claim
 from kmultiversx.utils import GENERATED_TOP_CELL
 
 if TYPE_CHECKING:
     from pyk.ktool.krun import KRun
 
 
-def test_generate_cterms_init_config(foundry_llvm_krun: KRun) -> None:
+def test_generate_cterms_init_config(kasmer_llvm_krun: KRun) -> None:
     # Given
-    conf = foundry_llvm_krun.definition.init_config(GENERATED_TOP_CELL)
+    conf = kasmer_llvm_krun.definition.init_config(GENERATED_TOP_CELL)
 
     # Check that the definition can be parsed
-    assert foundry_llvm_krun.kompiled_kore.definition
+    assert kasmer_llvm_krun.kompiled_kore.definition
 
     sym_conf, subst = split_config_from(conf)
 
@@ -24,5 +24,5 @@ def test_generate_cterms_init_config(foundry_llvm_krun: KRun) -> None:
     _, lhs, rhs = generate_claim('foo', ('BigUint', 'u32'), sym_conf, subst)
 
     # Then
-    foundry_llvm_krun.kast_to_kore(lhs.kast, GENERATED_TOP_CELL)
-    foundry_llvm_krun.kast_to_kore(rhs.kast, GENERATED_TOP_CELL)
+    kasmer_llvm_krun.kast_to_kore(lhs.kast, GENERATED_TOP_CELL)
+    kasmer_llvm_krun.kast_to_kore(rhs.kast, GENERATED_TOP_CELL)
